third_party/WebKit/LayoutTests/media/track/track-constants.html

<!DOCTYPE html>
<title>Test TextTrack constants</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
    assert_equals(HTMLTrackElement.NONE, 0);
    assert_equals(HTMLTrackElement.LOADING, 1);
    assert_equals(HTMLTrackElement.LOADED, 2);
    assert_equals(HTMLTrackElement.ERROR, 3);
});
</script>
